backend/requirements.txt
flask
PyMuPDF
pdfplumber
pdfminer.six